web_dashboard/backend/requirements_enhanced.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-dotenv==1.0.0
eventlet==0.33.3